Moving all Field.java inner classes to their own files. Split FieldTest.java to match. Also replaced SimpleArrayMap in FieldDropdown to preserve ordering.
