+++ blocklytest/src/androidTest/java/com/google/blockly/model/BlockTest.java
@@ -171,7 +171,7 @@ public class BlockTest extends AndroidTestCase {
         // Fields
         loaded = parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "8",
                 BlockTestStrings.FIELD_HAS_NAME), bf);
-        assertEquals("item", ((Field.FieldInput) loaded.getFieldByName("text_input")).getText());
+        assertEquals("item", ((FieldInput) loaded.getFieldByName("text_input")).getText());
         // A missing or unknown field name isn't an error, it's just ignored.
         parseBlockFromXml(BlockTestStrings.assembleFrankenblock("block", "9",
                 BlockTestStrings.FIELD_MISSING_NAME), bf);
