+++ blocklytest/src/androidTest/java/com/google/blockly/model/FieldVariableTest.java
@@ -1,0 +1,39 @@
+/*
+ *  Copyright 2016 Google Inc. All Rights Reserved.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package com.google.blockly.model;
+
+import android.test.AndroidTestCase;
+
+/**
+ * Tests for {@link FieldVariable}.
+ */
+public class FieldVariableTest extends AndroidTestCase {
+    public void testFieldVariable() {
+        FieldVariable field = new FieldVariable("fname", "varName");
+        assertEquals(Field.TYPE_VARIABLE, field.getType());
+        assertEquals("fname", field.getName());
+        assertEquals("varName", field.getVariable());
+
+        field.setVariable("newVar");
+        assertEquals("newVar", field.getVariable());
+
+        // xml parsing
+        assertTrue(field.setFromString("newestVar"));
+        assertEquals("newestVar", field.getVariable());
+        assertFalse(field.setFromString(""));
+
+        assertNotSame(field, field.clone());
+    }
+}
